requirement.txt
Django==2.1.15
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-versatileimagefield==2.0
django-cors-headers==3.2.1
auditlog==0.0.5
numpy==1.18.4
django-pyodbc-azure-2019==2.1.0.0
pandas==1.0.3
scipy==1.4.1
xlrd==1.2.0
scikit-learn==0.23.0
gower==0.0.5